requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
langchain-text-splitters>=0.0.1
faiss-cpu>=1.9.0
pymongo>=4.6.0
pypdf2>=3.0.0
python-docx>=1.1.0
tiktoken>=0.5.0
openai>=1.12.0
streamlit>=1.30.0
requests>=2.31.0
